Commiting changes, added Youtube & Comments
