+++ src/containers/Resource/Resource.js
