+++ src/containers/ShareResource/ShareResource.js
